+++ source/production/F0.Templates/templates/Roslyn-Generator/Roslyn.Generator.UnitTests/HelloWorldGeneratorUnitTests.cs
@@ -23,6 +23,7 @@ namespace Roslyn.Generator.UnitTests
 
 namespace Roslyn.Generated
 {
+	[global::System.CodeDom.Compiler.GeneratedCodeAttribute(""Roslyn.Generator"", ""1.0.0.0"")]
 	[global::System.AttributeUsage(global::System.AttributeTargets.Method, AllowMultiple = false)]
 	internal sealed class HelloWorldAttribute : global::System.Attribute
 	{